Add esbuild Refactor
