+++ apps/worker/package.json
@@ -16,12 +16,14 @@
   "packageManager": "pnpm@10.12.1",
   "devDependencies": {
     "@shellicar-reference-enterprise/typescript-config": "workspace:^",
+    "esbuild": "^0.25.5",
     "tsx": "^4.20.3",
     "vitest": "^3.2.4"
   },
   "dependencies": {
     "@hono/node-server": "^1.14.4",
     "@shellicar/build-version": "^1.1.0",
+    "@shellicar/core-di": "^2.4.0",
     "@temporalio/worker": "^1.11.8",
     "hono": "^4.8.2"
   }
