+++ apps/worker/src/core/di/modules.ts
@@ -1,0 +1,5 @@
+import type { IServiceModule } from '@shellicar/core-di';
+import { HonoServerModule } from '../http/HonoServerModule';
+import { TemporalWorkerModule } from '../temporal-worker/TemporalWorkerModule';
+
+export const modules = [HonoServerModule, TemporalWorkerModule] satisfies (typeof IServiceModule)[];
