+++ apps/worker/src/core/di/serviceProvider.ts
@@ -1,0 +1,6 @@
+import { createServiceCollection } from '@shellicar/core-di';
+import { modules } from './modules';
+
+const services = createServiceCollection();
+services.registerModules(...modules);
+export const serviceProvider = services.buildProvider();
