+++ apps/worker/src/core/http/HonoServerModule.ts
@@ -1,0 +1,9 @@
+import type { IServiceCollection, IServiceModule } from '@shellicar/core-di';
+import { IWorkerModule } from '../../interfaces';
+import { HonoServer } from './HonoServer';
+
+export class HonoServerModule implements IServiceModule {
+  registerServices(services: IServiceCollection): void {
+    services.register(IWorkerModule).to(HonoServer);
+  }
+}
